refactor(gateway): extract server_constructors.go from server.go
Phase 3b file split: moves NewServer, NewServerWithK8sClient, newAuthMiddleware,
ServerTestDeps, NewServerForTesting, NewServerWithMetrics, serverClients, and
createServerWithClients into their own file. Pure structural move: 154
Ginkgo specs still pass.
